chore: schnorr cleanup
